+++ src/lib/browser/BrowserTree.ts
@@ -6,9 +6,9 @@ import PQueue from 'p-queue'
 import Account from '../Account'
 import { Bookmark, Folder, ItemLocation, ItemType } from '../Tree'
 import Ordering from '../interfaces/Ordering'
-import url from 'url'
 import random from 'random'
 import seedrandom from 'seedrandom'
+import { isVivaldi } from './BrowserDetection'
 
 let absoluteRoot: {id: string}
 
@@ -29,6 +29,7 @@ export default class BrowserTree implements IResource {
   }
 
   async getBookmarksTree():Promise<Folder> {
+    const isVivaldiBrowser = await isVivaldi()
     const [tree] = await browser.bookmarks.getSubTree(this.rootId)
     await this.absoluteRootPromise
     const allAccounts = await (await Account.getAccountClass()).getAllAccounts()
@@ -43,7 +44,7 @@ export default class BrowserTree implements IResource {
         return
       }
       let overrideTitle, isRoot
-      if (node.parentId === this.absoluteRoot.id) {
+      if (node.parentId === this.absoluteRoot.id && !isVivaldiBrowser) {
         switch (node.id) {
           case '1': // Chrome
           case 'toolbar_____': // Firefox
@@ -121,7 +122,7 @@ export default class BrowserTree implements IResource {
       return
     }
     try {
-      if (self.location.protocol === 'moz-extension:' && url.parse(bookmark.url).hostname === 'separator.floccus.org') {
+      if (self.location.protocol === 'moz-extension:' && new URL(bookmark.url).hostname === 'separator.floccus.org') {
         const node = await this.queue.add(async() => {
           Logger.log('(local)CREATE: executing create ', bookmark)
           return browser.bookmarks.create({
@@ -152,7 +153,7 @@ export default class BrowserTree implements IResource {
       return
     }
     try {
-      if (self.location.protocol === 'moz-extension:' && url.parse(bookmark.url).hostname === 'separator.floccus.org') {
+      if (self.location.protocol === 'moz-extension:' && new URL(bookmark.url).hostname === 'separator.floccus.org') {
         // noop
       } else {
         await this.queue.add(async() => {
@@ -348,5 +349,9 @@ export default class BrowserTree implements IResource {
       absoluteRoot = (await browser.bookmarks.getTree())[0]
     }
     return absoluteRoot
+  }
+
+  isAvailable(): Promise<boolean> {
+    return Promise.resolve(true)
   }
 }