+++ src/lib/isTest.ts
@@ -1,0 +1,1 @@
+export const isTest = (new URL(window.location.href)).pathname.includes('test')