+++ src/test/test.js
@@ -144,13 +144,13 @@ describe('Floccus', function() {
     },
     {
       type: 'google-drive',
-      bookmark_file: random.float() + '.xbel',
+      bookmark_file: Math.random() + '.xbel',
       password: '',
       refreshToken: CREDENTIALS.password,
     },
     {
       type: 'google-drive',
-      bookmark_file: random.float() + '.xbel',
+      bookmark_file: Math.random() + '.xbel',
       password: random.float(),
       refreshToken: CREDENTIALS.password,
     },
@@ -2250,7 +2250,7 @@ describe('Floccus', function() {
               bookmark.parentId = serverTree.children.find(folder => folder.title !== 'foo').id
               const fooFolder = serverTree.children.find(folder => folder.title === 'foo')
               await adapter.updateBookmark(new Bookmark(bookmark))
-              // toLowerCase to accomodate chrome (since we normalize the title)
+              // toLowerCase to accommodate chrome (since we normalize the title)
               const secondBookmark = serverTree.children.find(folder => folder.title.toLowerCase() === secondBookmarkFolderTitle.toLowerCase()).children.find(item => item.type === 'bookmark')
               secondBookmark.parentId = fooFolder.id
               await adapter.updateBookmark(secondBookmark)
@@ -2725,7 +2725,7 @@ describe('Floccus', function() {
                 bookmark.parentId = serverTree.children.find(folder => folder.title !== 'foo').id
                 const fooFolder = serverTree.children.find(folder => folder.title === 'foo')
                 await adapter.updateBookmark(new Bookmark(bookmark))
-                // toLowerCase to accomodate chrome (since we normalize the title)
+                // toLowerCase to accommodate chrome (since we normalize the title)
                 const secondBookmark = serverTree.children.find(folder => folder.title.toLowerCase() === secondBookmarkFolderTitle.toLowerCase()).children.find(item => item.type === 'bookmark')
                 secondBookmark.parentId = fooFolder.id
                 await adapter.updateBookmark(secondBookmark)
@@ -4814,7 +4814,7 @@ describe('Floccus', function() {
         const setInterrupt = () => {
           if (!timeouts.length) {
             timeouts = new Array(1000).fill(0).map(() =>
-              ACCOUNT_DATA.type === 'nextcloud-bookmarks' ? random.int(50000, 150000) : random.int(1000,30000)
+              ACCOUNT_DATA.type === 'nextcloud-bookmarks' ? random.int(50000, 150000) : random.int(100,3000)
             )
           }
           const timeout = timeouts[(i++) % 1000]
@@ -4841,10 +4841,29 @@ describe('Floccus', function() {
           await account2.init()
 
           if (ACCOUNT_DATA.type === 'fake') {
-            // Wrire both accounts to the same fake db
-            account2.server.bookmarksCache = account1.server.bookmarksCache = new Folder(
+            // Wire both accounts to the same fake db
+            // We do not set the cache properties to the same object, because we want to only write onSynComplete
+            let fakeServerDb = new Folder(
               { id: '', title: 'root', location: 'Server' }
             )
+            account1.server.bookmarksCache = new Folder(
+              { id: '', title: 'root', location: 'Server' }
+            )
+            account2.server.bookmarksCache = new Folder(
+              { id: '', title: 'root', location: 'Server' }
+            )
+            account1.server.onSyncStart = () => {
+              account1.server.bookmarksCache = fakeServerDb.clone(false)
+            }
+            account1.server.onSyncComplete = () => {
+              fakeServerDb = account1.server.bookmarksCache.clone(false)
+            }
+            account2.server.onSyncStart = () => {
+              account2.server.bookmarksCache = fakeServerDb.clone(false)
+            }
+            account2.server.onSyncComplete = () => {
+              fakeServerDb = account2.server.bookmarksCache.clone(false)
+            }
             account2.server.__defineSetter__('highestId', (id) => {
               account1.server.highestId = id
             })
@@ -5056,7 +5075,7 @@ describe('Floccus', function() {
           tree1AfterFinalSync = null
         })
 
-        it('should handle fuzzed changes', async function() {
+        it('should handle fuzzed changes from one client', async function() {
           const localRoot = account1.getData().localRoot
           let bookmarks = []
           let folders = []
@@ -5632,7 +5651,7 @@ describe('Floccus', function() {
             await randomlyManipulateTreeWithDeletions(account1, folders1, bookmarks1, 35)
             await randomlyManipulateTreeWithDeletions(account2, folders2, bookmarks2, 35)
 
-            console.log(' acc1: Moved items')
+            console.log(' acc1&acc2: Moved items')
 
             let tree1BeforeSync = await account1.localTree.getBookmarksTree(
               true
@@ -5749,11 +5768,8 @@ describe('Floccus', function() {
             serverTreeAfterInit = null
           }
         })
-
-        it.skip('should handle fuzzed changes with deletions from two clients with interrupts', async function() {
-          if (ACCOUNT_DATA.type === 'nextcloud-bookmarks' && ACCOUNT_DATA.oldAPIs) {
-            return this.skip()
-          }
+        let interruptBenchmark
+        it('should handle fuzzed changes with deletions from two clients with interrupts' + (ACCOUNT_DATA.type === 'fake' ? ' (with caching)' : ''), interruptBenchmark = async function() {
           const localRoot = account1.getData().localRoot
           let bookmarks1 = []
           let folders1 = []
@@ -5872,7 +5888,7 @@ describe('Floccus', function() {
             await randomlyManipulateTreeWithDeletions(account1, folders1, bookmarks1, 35)
             await randomlyManipulateTreeWithDeletions(account2, folders2, bookmarks2, 35)
 
-            console.log(' acc1: Moved items')
+            console.log(' acc1 &acc2: Moved items')
 
             let tree1BeforeSync = await account1.localTree.getBookmarksTree(
               true
@@ -5984,6 +6000,21 @@ describe('Floccus', function() {
             serverTreeAfterInit = null
           }
         })
+
+        if (ACCOUNT_DATA.type === 'fake') {
+          it('should handle fuzzed changes with deletions from two clients with interrupts (no caching adapter)', async function() {
+            // Wire both accounts to the same fake db
+            // We set the cache properties to the same object, because we want to simulate nextcloud-bookmarks
+            account1.server.bookmarksCache = account2.server.bookmarksCache = new Folder(
+              { id: '', title: 'root', location: 'Server' }
+            )
+            delete account1.server.onSyncStart
+            delete account1.server.onSyncComplete
+            delete account2.server.onSyncStart
+            delete account2.server.onSyncComplete
+            await interruptBenchmark()
+          })
+        }
 
         it('unidirectional should handle fuzzed changes from two clients', async function() {
           await account2.setData({...account2.getData(), strategy: 'slave'})
@@ -6658,7 +6689,7 @@ async function syncAccountWithInterrupts(account) {
   try {
     expect(account.getData().error).to.not.be.ok
   } catch (e) {
-    if (!account.getData().error.includes('E027')) {
+    if (!account.getData().error.includes('E026')) {
       throw e
     } else {
       console.log(account.getData().error)