+++ src/ui/views/Update.vue
@@ -85,7 +85,17 @@ export default {
           href: 'https://github.com/users/marcelklehr/sponsorship',
           label: this.t('LabelGithubsponsors'),
           description: this.t('DescriptionGithubsponsors')
-        }
+        },
+        {
+          href: 'https://www.patreon.com/marcelklehr',
+          label: this.t('LabelPatreon'),
+          description: this.t('DescriptionPatreon')
+        },
+        {
+          href: 'https://www.ko-fi.com/marcelklehr',
+          label: this.t('LabelKofi'),
+          description: this.t('DescriptionKofi')
+        },
       ]
     }
   },